requirements.txt
Django==4.2.6
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0 